imdb matching - add exponential factor to score
for movies like LOTR, the filename is often the last part, e.g. "the fellowship of the ring", and in such case, there are other matches.
but, we know that the imdb query brings the most relevant result first, so we can use that to add an increasing factor as we examine later results